- remove some warnings about variable resizing with possible loss of information
git-svn-id: https://svn.openxp.de/openxp/branches/branch_3_7_8@6898 65356fb5-50e6-0310-a06b-fbc0adf8d39c

+++ fidoglob.pas
@@ -145,7 +145,7 @@ procedure TNodeList.LoadConfigFromFile;       { NODELST.CFG laden }
 var t     : text;
     s     : string;
     ss    : string[20];
-    p     : byte;
+    p     : Integer;
     fa    : fidoadr;
     NlItem: TNodeListItem;
 begin
@@ -225,7 +225,8 @@ begin
 end;
 
 function  TNodeList.GetFileName(n:integer):string;
-var p : byte;
+var
+  p: Integer;
 begin
   if n>=Count then
     result:=''
