+++ help.pas
@@ -48,7 +48,7 @@ const maxpages = 2000;
       HHeadNotHigh   = false;
 
 procedure sethelpcol(col,colhi,colqvw,colselqvw:byte);
-function  inithelp(name:string; xh,yh:byte;
+function  inithelp(name:string; xh,yh: Integer;
                    invers,blocksatz,headline:boolean):boolean;
 procedure sethelppos(_x,_y,height:word);
 procedure help_printable(printchar:taste; pinit,pexit:string);
@@ -165,7 +165,7 @@ begin
 end;
 
 
-function inithelp(name:string; xh,yh:byte;
+function inithelp(name:string; xh,yh: Integer;
                   invers,blocksatz,headline:boolean):boolean;
 
 var ixadr : longint;
